Updated chaching to include tags for our non-stopword lemmas.
Now the word embedding visualization runs on the cache very fast!
One problem: the words and tags in fasttext.py are not the same length. This is
definitely because I'm saving the lemmas as a STRING in the cache, and the tags as a list.
Originally I saved the lemmas as a string because this was necessary for LSA and LDA.
But I will go back and change it so that the non-stopword lemmas are still a list,
and then add the extra step to LSA to LDA to convert lemmas to string,
that way we don't get non-NounPhrases in the embeddings visualization.