Sorted heatmap by year...
But now there's issues with Plotly, getting the xaxis labels to plot.
Plotly deletes duplicate x-axis labels, so I'm trying to force them in the graph with "xaxis" in conge_heatmap2.
This didn't work so I suspected the library was behind, so I did a bower update.
The bower update updated Plotly to 1.29.x, and the venn_diagram library, and some others. Hopefully this doesn't cause problems...