Fixed heatmap x-axis so that there are no repeats labels.
This way Plotly doesn't delete the labels.
I could not force Plotly to accept duplicate labels by updating the "layout" in the Javascript, so instead I updated the labels themselves.
So instead if there are duplicate YEAR, AUTHOR labels (because an author published a lot that year!!), the code in nes.py checks for duplicates and changes the label to be "YEARb, AUTHOR" for second, "YEARc, AUTHOR" for third, etc etc.
Better hope no author has more than 26 publications in a year heheh.