Updating coge statistics to reference the db more.
This includes writing a new function in database_management, and tweaking an existing function in content_management that calls it.
The "queries" db table is not well populated, so current CoGe results says "None" for several things.
So I will need to populate the queries table and integrate that into @app.route(/results) for sure.