Updated coge statistics bar chart to be stacked.
This workeed easily for the coge papers, however if 2 papers don't have the same year_range, then there will be problems.
Will need to update the bar chart function to handle this, and make the stacked barchart for @app.route(/results).

Now should be ready to start integrating changes on the /results side :)