Journal vis updates only when needed.
That being when there is a brand new query, OR when there are new citations to an existing query.
If there are new citations, it force updates the journals vis. Otherwise it just returns the chache one.

Alsi in journalvis.py, I had to update the regex to deal with insane people who put the date as "20170607" for the metadata on PubMed(Central).