TRYING (and failing) to get Embeddings in /results.
For some reason when i try to create sub directories with os.makedirs(), it is saying that these sub directories exist...
But when I go check, they do not exist.
Possibly there is a race condition going on (according to the Internet) but I'm unsure how that would happen here since its not being run in parallel at the moment.

What is happening?!?!?!?! :'(