Topic model cache checks are better!!
Even if the query is new, or the query is updated, it will check when the json cache was last updated.
This way if someone runs an analysis on a new query, it will load the cache if it was modified recently enough, so that the analysis doesn't always run from scratch.