Journals and Stats: "None" date.
Exception handling for if the publication date on Entrez is "None".
Had this happen with PMID: 26220524, where Entrez had no date for 2 of the papers.
If the date is "None", then I re-set it to be 2018 (the future!!!!).
I know this is hacky... this will need a better solution in the future but is low priority right now.