journals bugfix
If there were no new updates for query documents, and both had been seen by the database, it always tries to load the file. But if there is no file for some reason, it didn't create it. Changed it to check for the file and create it if there isn't one.