fixed missing text retrieval.
re-retrieves any missing txts or annotated jsons that are in the db but missing from filesystem.
also updates the lemma and nes samples cache for that item.
if we had to re-retrieve ANY papers that the db says exits, then the nes and lemma samples cache for the collection of citations for the query is also re-generated!